// src/app/business-logic/model/projects/projectsCreateRequest.ts
/**
 * projects
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.12
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface ProjectsCreateRequest {
    /**
     * Project name Unique within the company.
     */
    name: string;
    /**
     * Project description.
     */
    description: string;
    /**
     * User-defined tags
     */
    tags?: Array<string>;
    /**
     * System tags. This field is reserved for system use, please don't use it.
     */
    system_tags?: Array<string>;
    /**
     * The default output destination URL for new tasks under this project
     */
    default_output_destination?: string;
}